+++ NEWS.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Since PR[#XXXX](https://github.com/gridap/Gridap.jl/pull/XXXX) (number to be filled when the PR is opened).
-  - New variants (accessible via kwarg `flavor=:BMM`) of the `BarycentricPΛBasis` and `BarycentricPmΛBasis` bases of the full and trimmed polynomial PΛ¹ spaces, that admit closed-form change-of-basis with respect to relabelling/permutation of the simplex vertices.
+  - New variants (accessible via kwarg `flavor=:BMM`) of the `BarycentricPΛBasis` and `BarycentricPmΛBasis` bases of the PᵣrΛᵏ and P⁻ᵣrΛᵏ spaces. For form order k=1, both flavors of both bases implement a closed-form change-of-basis with respect to relabelling/permutation of the simplex vertices.
   - Added the associated reference FEs (named `rotating_pλ`, `trimmed_pλ`) which provide H(curl)-conforming FE spaces on meshes with arbitrary simplex ordering (`NonOriented`).
 
 ## [0.20.9] - 2026-08-17